+++ src/main/java/com/mwt/consumers/ConsumePolicies.java
@@ -1,0 +1,9 @@
+package com.mwt.consumers;
+
+import com.mwt.policies.Policy;
+
+import java.util.List;
+
+public interface ConsumePolicies<T> {
+  void updatePolicy(List<Policy<T>> newPolicies);
+}
