+++ src/main/java/com/mwt/consumers/ConsumePolicy.java
@@ -1,0 +1,7 @@
+package com.mwt.consumers;
+
+import com.mwt.policies.Policy;
+
+public interface ConsumePolicy<T> {
+  void updatePolicy(Policy<T> newPolicy);
+}
