+++ src/main/java/com/mwt/consumers/ConsumeScorer.java
@@ -1,0 +1,7 @@
+package com.mwt.consumers;
+
+import com.mwt.scorers.Scorer;
+
+public interface ConsumeScorer<T> {
+  void updateScorer(Scorer<T> newScorer);
+}
